Re-organize into a Maven structure.
This commit re-organizes the Spark code base into a Apache Maven-compatible struture.

The project in this commit can be used to create a working Spark instance, by executing:

    mvn clean package

An assembly (tied to the package phase) is used to generate a distributable directory.

The includes EJ-Technologies install4j configuration file has been adjusted to reflect the changes, and can be used to generate most types of distributables. This requires the non-free install4j application to be installed at the host that executes the build.

The build currently misses all 'external' plugins. Work remains to be done to move that into a Maven structure. A future improvement to this structure could include splitting off the assembly part from the core module, into a dedicated module, that allows for easy inclusiong of plugin-module artifacts.